api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.4.2
base58==2.1.1
openai==1.57.3
python-dotenv==1.0.0
redis==5.2.1